# openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise E-Commerce API
  version: 0.1.0
  description: REST API for catalog, checkout, orders, and payments.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/admin/inventory:
    get:
      summary: Admin list inventory items
      operationId: adminListInventory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: low_stock
          in: query
          schema:
            type: boolean
            default: false
        - name: q
          in: query
          description: Filter by SKU or product name (case-insensitive)
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Inventory list with variant SKU and availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInventoryListResponse"
  /api/v1/admin/inventory/{variant_id}:
    patch:
      summary: Admin adjust inventory on-hand quantity
      operationId: adminAdjustInventory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: variant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAdjustInventoryRequest"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInventoryItemSchema"
        "409":
          description: Version conflict
        "422":
          description: on_hand below reserved quantity
...
  /api/v1/admin/integrations/moysklad/status:
    get:
      summary: MoySklad integration status
      operationId: adminMoySkladStatus
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Integration status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladIntegrationStatusResponse"
        "401":
          description: Unauthorized
  /api/v1/admin/integrations/moysklad/sync/pull:
    post:
      summary: Pull catalog and stock from MoySklad
      operationId: adminMoySkladPullCatalog
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladPullResponse"
        "503":
          description: MoySklad not configured
  /api/v1/admin/integrations/moysklad/sync/stock:
    post:
      summary: Pull stock from MoySklad
      operationId: adminMoySkladPullStock
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stock pull result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladStockPullResponse"
        "503":
          description: MoySklad not configured
  /api/v1/admin/integrations/moysklad/sync/resync:
    post:
      summary: Full resync (catalog, stock, pending order exports)
      operationId: adminMoySkladFullResync
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Full resync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladFullResyncResponse"
        "503":
          description: MoySklad not configured
  /api/v1/admin/integrations/moysklad/sync/trigger:
    post:
      summary: "Deprecated: use sync/pull"
      operationId: adminMoySkladTriggerSync
      deprecated: true
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladPullResponse"
  /api/v1/admin/integrations/moysklad/webhooks/enabled:
    patch:
      summary: Enable or disable MoySklad webhook processing
      operationId: adminMoySkladWebhooksEnabled
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoySkladWebhooksEnabledRequest"
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks_enabled:
                    type: boolean
...
  /api/v1/admin/integrations/moysklad/logs:
    get:
      summary: MoySklad sync logs
      operationId: adminMoySkladSyncLogs
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent sync logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladSyncLogListResponse"
  /api/v1/admin/integrations/moysklad/orders/{order_number}/export:
    post:
      summary: Export order to MoySklad
      operationId: adminMoySkladExportOrder
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: order_number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoySkladOrderExportResponse"
        "404":
          description: Order not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminInventoryItemSchema:
      type: object
      required:
        - variant_id
        - sku
        - product_name
        - sync_source
        - quantity_on_hand
        - quantity_reserved
        - available
        - version
        - is_low_stock
      properties:
        variant_id:
          type: string
          format: uuid
        sku:
          type: string
        product_name:
          type: string
        sync_source:
          type: string
          enum: [manual, moysklad]
        quantity_on_hand:
          type: integer
        quantity_reserved:
          type: integer
        available:
          type: integer
        version:
          type: integer
        is_low_stock:
          type: boolean
    AdminInventoryListResponse:
      type: object
      required: [items, total, page, limit, low_stock_threshold]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminInventoryItemSchema"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        low_stock_threshold:
          type: integer
    AdminAdjustInventoryRequest:
      type: object
      required: [quantity_on_hand, reason, version]
      properties:
        quantity_on_hand:
          type: integer
          minimum: 0
        reason:
          type: string
          enum: [restock, damage, correction, return]
        version:
          type: integer
          minimum: 0
...
    MoySkladIntegrationStatusResponse:
      type: object
      required:
        - configured
        - read_only
        - order_export_enabled
        - webhooks_enabled
        - pending_order_exports
        - errors_last_24h
      properties:
        configured:
          type: boolean
        read_only:
          type: boolean
          default: true
        order_export_enabled:
          type: boolean
        store_id:
          type: string
          nullable: true
        webhooks_enabled:
          type: boolean
        pending_order_exports:
          type: integer
        last_full_sync_at:
          type: string
          format: date-time
          nullable: true
        last_incremental_sync_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
        errors_last_24h:
          type: integer
    MoySkladPullResponse:
      type: object
      required: [status, direction]
      properties:
        status:
          type: string
        direction:
          type: string
          default: inbound
        products_created:
          type: integer
        products_updated:
          type: integer
        variants_created:
          type: integer
        variants_updated:
          type: integer
        stock_rows_applied:
          type: integer
        errors:
          type: array
          items:
            type: string
    MoySkladStockPullResponse:
      type: object
      required: [status, direction, stock_rows_applied]
      properties:
        status:
          type: string
        direction:
          type: string
        stock_rows_applied:
          type: integer
        errors:
          type: array
          items:
            type: string
    MoySkladFullResyncResponse:
      type: object
      required: [status, direction]
      properties:
        status:
          type: string
        direction:
          type: string
        products_created:
          type: integer
        products_updated:
          type: integer
        variants_created:
          type: integer
        variants_updated:
          type: integer
        stock_rows_applied:
          type: integer
        order_exports_succeeded:
          type: integer
        errors:
          type: array
          items:
            type: string
    MoySkladWebhooksEnabledRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
...
    MoySkladSyncLogSchema:
      type: object
      required: [id, direction, entity_type, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        direction:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
          nullable: true
        status:
          type: string
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    MoySkladSyncLogListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MoySkladSyncLogSchema"
    MoySkladOrderExportResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
        moysklad_order_id:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
...